requirements.txt
pypdf
paddleocr
paddlepaddle
pdf2image
Pillow
sentence-transformers
faiss-cpu
numpy
openai
streamlit
